+++ packages/storage-plugin/src/engines.ts
@@ -7,7 +7,6 @@ declare const ngServerMode: boolean;
 export const LOCAL_STORAGE_ENGINE = /* @__PURE__ */ new InjectionToken<StorageEngine | null>(
   typeof ngDevMode !== 'undefined' && ngDevMode ? 'LOCAL_STORAGE_ENGINE' : '',
   {
-    providedIn: 'root',
     factory: () => (typeof ngServerMode !== 'undefined' && ngServerMode ? null : localStorage)
   }
 );
@@ -15,7 +14,6 @@ export const LOCAL_STORAGE_ENGINE = /* @__PURE__ */ new InjectionToken<StorageEn
 export const SESSION_STORAGE_ENGINE = /* @__PURE__ */ new InjectionToken<StorageEngine | null>(
   typeof ngDevMode !== 'undefined' && ngDevMode ? 'SESSION_STORAGE_ENGINE' : '',
   {
-    providedIn: 'root',
     factory: () =>
       typeof ngServerMode !== 'undefined' && ngServerMode ? null : sessionStorage
   }